Isolate set_arm_clock into its own module
